% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee.R
\docType{data}
\name{bee_surveys}
\alias{bee_surveys}
\title{Bee surveys}
\format{A data.frame with the following columns:
\itemize{
  \item date: Date, date of survey
  \item round: character, round number for the survey: 1,2,3
  \item observer: character, name of surveyor
  \item transectID: character, id of the transect
  \item transect_length: integer, length of transect in meters
}}
\usage{
bee_surveys
}
\description{
A list of all surveys conducted for bees. This was primarily used to complete
the \code{\link{bee}} with zeros.
}
\keyword{datasets}
