% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daubenmire.R
\docType{data}
\name{daubenmire}
\alias{daubenmire}
\title{Daubenmire}
\format{A data frame with the following variables:
\itemize{
  \item daubenmire: daubenmire land cover %'s, count data for 3 milkweed species, and litter depth in cm observed at 10m intervals inside a 0.5m square
  \item year: year of observation
  \item month: month of observation
  \item day: day of observation
  \item recorder: individual who observed (and possibly recorded) the observations
  \item siteID: id of the site
  \item transectID: id of the transect
  \item round: the round number: 1, 2, or 3
  \item distance: the interval of the observation in length of meters along transect
  \item landcover
  \item percentage
  \item section
}}
\usage{
daubenmire
}
\description{
This data set contains daubenmire land cover %'s, count data for 3 milkweed species, and litter depth in cm observed at 10m intervals inside a 0.5m square along a transect.
}
\keyword{datasets}
