% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daubenmire.R
\docType{data}
\name{daubenmire_raw}
\alias{daubenmire_raw}
\title{Daubenmire raw data}
\format{A data.frame with the following columns:
\itemize{
  \item filepath: character, the name of the file that these data came from
  \item date: Date, the date survey was taken
  \item round: character, round number for the survey: 1, 2, or 3
  \item observer: character, the individual/group who did the survey
  \item siteID: character, unique identifier for the site
  \item transectID: character, unique identified for the transect
  \item section: character, point along the transect at which data was collected inside a 0.5m square
  \item remaining columns are cover or counts for various responses
}}
\usage{
daubenmire_raw
}
\description{
Contains raw daubenmire data. For analysis, you likely do not want this data
frame, but instead want one of these three:
\code{\link{cover}}, \code{\link{milkweed}}, \code{\link{litter}}.
}
\details{
This data.frame exists so that data issues can be checked. If you find an
error in one of the three data.frames listed above and the error is also in
this data.frame, then the error is in the raw data. If you find an error in
one of the three data.frames listed above and the error is not here, then 
it is caused by the script data-raw/daubenmire.R that constructs those three
data.frames.
}
\seealso{
\code{\link{cover}}, \code{\link{milkweed}}, \code{\link{litter}}, \code{\link{daubenmire_surveys}}
}
\keyword{datasets}
