% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daubenmire.R
\docType{data}
\name{daubenmire_surveys}
\alias{daubenmire_surveys}
\title{Daubenmire surveys}
\format{A data.frame with the following columns:
\itemize{
  \item date: Date, survey date
  \item round: character, round number for the survey: 1, 2, or 3
  \item observer: character, the individual/group who did the survey
  \item transectID: character, unique identified for the transect
}}
\usage{
daubenmire_surveys
}
\description{
Dates and observers for Daubenmire surveys
}
\details{
This data.frame was used primarily to fill in any missing data in the 
Daubenmire data sets. As of 2020-05-15, there was no missing data and thus
this is just duplicated information for what can be found in the other 
Daubenmire data.frames: 
\code{\link{cover}}, \code{\link{milkweed}}, \code{\link{litter}}, \code{\link{daubenmire_raw}}
}
\seealso{
\code{\link{cover}}, \code{\link{milkweed}}, \code{\link{litter}}
}
\keyword{datasets}
