% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\docType{data}
\name{environment}
\alias{environment}
\title{Environment}
\format{A data frame with the following variables:
\itemize{
  \item year: year of observation
  \item month: month of observation
  \item day: day of observation
  \item siteID: id of the site
  \item transectID: id of the transect
  \item round: the round number: 1, 2, or 3
  \item currently_flowering_plants: daubenmire style land cover of currently flowering plants in defined area surrounding plot. Grouped by percentages: 0,0-5,5-25,25-50,50-75,75-100
  \item dominant_flowering_species: dominant flowering plant species in defined area surrounding plot
  \item milkweed_ramet: count of milkweed ramets in defined area surrounding plot
  \item sky: cloud cover group name(clear, mostly clear, mostly cloudy, cloudy, rain)
  \item temperature: temperature in fahrenheit
  \item wind: wind speed in miles per hour
}}
\usage{
environment
}
\description{
This data set contains weather data and floral/milkweed resources of surrounding landscape (milkweed ramet counts, % of flowering plants, and dominant flowering species) at the time of observation
No environment data exists for 2016.
}
\details{
‘temperature’ °F. Measured with an anemometer on site.

‘wind’ Miles per hour. Measured with an anemometer on site.

‘sky’ 'Clear': no clouds. 'Mostly clear': <50% cloud cover. 'Mostly cloudy': >50% cloud cover. 'Cloudy': 100% cloud cover. 'Rain': rain

‘milkweed_ramet’ Count data of total milkweed ramets of any milkweed species observed within defined banded area around plot perimeter. Band was defined as area which was between 5-25 meters from the plot edges, essentially a ‘frame’ around the plot which had a width of 20 meters wide, starting 5m from the plot edge.

‘currently_flowering_plants’Visually estimated % of land cover which is covered by currently flowering plants within same defined area as ‘milkweed_ramet’ data collection. A 20 meter wide frame around plot exterior, starting 5m from the plot edge. In essence this data is attempting to do a daubenmire reading for % of cover of flowering plants across a large area.
‘0’: 0% (no flowering plants present in defined area)
‘0-5’: <0-5%
‘25-50’: 25-50%
‘50-75’: 50-75%
’75-100’: 75-100%
‘dominant_flowering_species’Visually estimated evaluation of which flowering species was present at greatest total number of plants inside the 20 meter wide frame around the plot exterior, starting 5m from the plot edge. Observation was recorded by its common name. Common names and corresponding scientific names are recorded in the file ‘nectar_species_scientific_name_counting_unit’ located in the folder 'data-raw/nectar' in the Git repository ISUmonarch.
If no flowering species were present (‘currently_flowering_plants’ was 0) no data for ‘dominant_flowering_species’ is recorded
If multiple species were codominant; multiple species were listed separated by a semicolon
}
\keyword{datasets}
