% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarch.R
\docType{data}
\name{monarch}
\alias{monarch}
\title{Monarch}
\format{A data frame with the following variables:
\itemize{
  \item year: year of observation
  \item month: month of observation
  \item day: day of observation
  \item recorder: individual who observed (and possibly recorded) the observations
  \item siteID: id of the site
  \item transectID: id of the transect
  \item round: the round number: 1, 2, or 3
  \item end_time: time survey ended
  \item sky: cloud cover (clear, mostly clear, mostly cloudy, cloudy, rain)
  \item start_time: time survey began
  \item temp: temperature in fahrenheit at start of survey
  \item wind: wind in miles per hour at start of survey
  \item distance: the interval of the observation in length of meters along transect
  \item count: count data for monarch adults, eggs, and instars. count data for palmer amaranth
}}
\usage{
monarch
}
\description{
This data set contains weather data and count data for monarch adults, eggs, and instars during a survey as well as count data for palmer amaranth.
}
\keyword{datasets}
