% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nectar.R
\docType{data}
\name{nectar_raw}
\alias{nectar_raw}
\title{Raw nectar and ramet data}
\format{A data.frame with the following columns:
\itemize{
  \item filepath: character, path to the raw data file
  \item date: Date, survey date
  \item round: character, the round number: 1, 2, or 3
  \item observer
  \item siteID
  \item transectID
  \item Nectar Plant Species
  \item the remaining columns are counts of the number of observations in 
  the corresponding section of the transect
}}
\usage{
nectar_raw
}
\description{
This data.frame contains the raw nectar and ramet data. The purpose of the 
data.frame is to allow investigation of data issues. If you find an issue
in the \code{\link{nectar}} or \code{\link{ramet}} data.frames and the issue
is also in this raw file, then the issue is with the raw data in 
data-raw/nectar. If you find an issue in the above data.frames, but the
issue is not in this raw file, then the issue arose in the processing script
data-raw/nectar. If no observations was made during a survey, there will be
no information in this data.frame, but the survey will appear in
\code{\link{nectar_surveys}}
}
\seealso{
\code{\link{nectar}}, \code{\link{ramet}}, \code{\link{nectar_surveys}}
}
\keyword{datasets}
