% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nectar.R
\docType{data}
\name{nectar_surveys}
\alias{nectar_surveys}
\title{Nectar and ramet surveys}
\format{A data.frame with the following columns:
\itemize{
  \item date:
  \item round:
  \item observer:
  \item transectID:
  \item section:
}}
\usage{
nectar_surveys
}
\description{
This data.frame contains all the nectar and ramet surveys that were conducted.
The primary purpose of this data.frame was to complete the 
\code{\link{nectar}} and \code{\link{ramet}} data.frames with missing zero 
counts.
}
\seealso{
\code{\link{nectar}}, \code{\link{ramet}}, \code{\link{nectar_raw}}
}
\keyword{datasets}
