% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nectar.R
\docType{data}
\name{ramet_strip}
\alias{ramet_strip}
\title{Ramet strip data}
\format{An object of class \code{data.frame} with 20 rows and 25 columns.}
\usage{
ramet_strip
}
\description{
It is unclear exactly what these data are. They are derived from data files
in data-raw/nectar, but they are NOT part of the formal survey.
}
\keyword{datasets}
