% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robel.R
\docType{data}
\name{robel}
\alias{robel}
\title{Robel}
\format{A data frame with the following variables:
\itemize{
  \item robel: visual obstruction robel pole readings observed at 10m intervals. # indicates # of 10cm bands visually obstructed at 4m distance from pole, 1m height.
  \item year: year of observation
  \item month: month of observation
  \item day: day of observation
  \item recorder: individual who observed (and possibly recorded) the observations
  \item siteID: id of the site
  \item transectID: id of the transect
  \item round: the round number: 1, 2, or 3
  \item distance: the interval of the observation in length of meters along transect
  \item count: robel pole reading. Numbers above 16 are 16. 16 is max.
  \item direction: direction from pole in which observation is taken. North, east, west, or south.
}}
\usage{
robel
}
\description{
This data set contains visual obstruction robel pole readings observed at 10m intervals along a transect.
}
\keyword{datasets}
