% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robel.R
\docType{data}
\name{robel_raw}
\alias{robel_raw}
\title{Raw robel data}
\format{A data.frame with the following columns:
\itemize{
  \item filepath: character, the path and filename for the original source data
  \item date:
  \item round:
  \item observer:
  \item siteID:
  \item transectID:
  \item Direction:
  \item the remaining columns are the locations (sections) where the 
  measurements were taken along the transect
}}
\usage{
robel_raw
}
\description{
Contains raw robel data and the filenames associated with the raw data. 
This is primarily meant to be used for data issues. If you find an error in
\code{\link{robel}} and the error is also here, then the error in the raw
data file. If you find an error in \code{\link{robel}} and the error is not
here, then the error was caused by the processing script data-raw/robel.R.
}
\seealso{
\code{\link{robel}}, \code{link{robel_surveys}}
}
\keyword{datasets}
