% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robel.R
\docType{data}
\name{robel_surveys}
\alias{robel_surveys}
\title{Robel surveys}
\format{A data.frame with the columns:
\itemize{
  \item date:
  \item round:
  \item transectID:
  \item observer:
}}
\usage{
robel_surveys
}
\description{
The list of transects where robel surveys were done
}
\details{
This data set is unlikely to be needed and really only provides the names
of observers.
}
\seealso{
\code{\link{robel}}, \code{\link{robel_raw}}
}
\keyword{datasets}
