% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect.R
\docType{data}
\name{transect}
\alias{transect}
\title{Transect}
\format{A data frame with the following variables:
\itemize{
  \item siteID: id of the site
  \item transectID: id of the transect
  \item grant: grant associated with each transect/site. 
  
  'iacig' is Iowa NRCS Consevation Innovation Grant. Establishing habitat in existing grasslands on top of saturated buffers, as riparian buffers, beside oxbows, and bordering wetlands.
  
  'ncig' is National NRCS Conservation Innovation Grant. Establishing habitat in underutilized agricultural grassland field edges.
 
  'soy' is Iowa Soybean Association Grant. Establishing habitat on top of constructed bioreactors.
  
  'pork' is Iowa Pork Producers Association/National Pork Board Grant. Replacing lawn with pollinator habitat at hog confinement facilities.
    
  \item county: county in Iowa associated with each transect/site.
}}
\usage{
transect
}
\description{
Data set contains corresponding grant and county for each transect/site.
}
\keyword{datasets}
