% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\docType{data}
\name{landscape2016}
\alias{landscape2016}
\title{Landscape}
\format{A data frame with the following variables:
\itemize{
  \item landscape: contains general information about landscape surrounding plots. Yes/no answers to milkweed presence, flowering plants, and habitat type 
  \item year: year of observation
  \item month: month of observation
  \item day: day of observation
  \item recorder: individual who observed (and possibly recorded) the observations
  \item siteID: id of the site
  \item transectID: id of the transect
  \item round: the round number: 1, 2, or 3
}}
\usage{
landscape2016
}
\description{
This data set contains general information about the landscape surrounding the plots
Data set was only used in 2016. Replaced by 'environment'
}
\keyword{datasets}
