% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nectar.R
\docType{data}
\name{nectar}
\alias{nectar}
\title{Nectar plant species}
\format{A data frame with the following variables:
\itemize{
  \item year: year of observation
  \item month: month of observation
  \item day: day of observation
  \item siteID: id of the site
  \item transectID: id of the transect
  \item round: the round number: 1, 2, or 3
  \item section: the interval of the observation in length of meters along transect
  \item Nectar Plant Species: common name for nectar plant species which were flowering at the time of the survey
  \item count: number of plant species specific floral units observed
}}
\usage{
nectar
}
\description{
This data set contains the count of the number of plant species specific 
floral units observed within certain sections of a transect.
}
\section{NewSection}{
 test

Here is some information.
}

\references{
Here is a reference.
}
\keyword{datasets}
