% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nectar}
\alias{nectar}
\title{Nectar plant species}
\format{A data frame with the following variables:
\itemize{
  \item species: common name for nectar plant species
  \item year: year of observation
  \item month: month of observation
  \item day: day of observation
  \item recorder: individual who recorded the observations
  \item site: location of observation
  \item transect: id of the transect
  \item round: the round number
  \item distance: the interval of the observation
  \item count: number of plants observed
}}
\usage{
nectar
}
\description{
This data set contains the count of the number of plants observed within certain distances of a transect.
}
\keyword{datasets}

